Switch everything to C++

+++ src/tank.cpp
@@ -11,8 +11,16 @@
 const uint8_t max_shells[] = {5, 1, 1, 1, 1, 3, 2, 5, 5, 2};
 const uint8_t max_mines[] = {2, 0, 0, 0, 4, 0, 0, 2, 2, 2};
 const uint8_t max_bounces[] = {1, 1, 1, 0, 1, 1, 2, 1, 1, 0};
-const uint8_t tank_velocities[] = {TANK_SPEED_NORMAL, 0, TANK_SPEED_SLOW, TANK_SPEED_SLOW, TANK_SPEED_HIGH,
-                                   TANK_SPEED_NORMAL, 0, TANK_SPEED_HIGH, TANK_SPEED_NORMAL, TANK_SPEED_BLACK};
+const uint8_t tank_velocities[] = {(uint8_t)TANK_SPEED_NORMAL,
+                                   (uint8_t)0,
+                                   (uint8_t)TANK_SPEED_SLOW,
+                                   (uint8_t)TANK_SPEED_SLOW,
+                                   (uint8_t)TANK_SPEED_HIGH,
+                                   (uint8_t)TANK_SPEED_NORMAL,
+                                   (uint8_t)0,
+                                   (uint8_t)TANK_SPEED_HIGH,
+                                   (uint8_t)TANK_SPEED_NORMAL,
+                                   (uint8_t)TANK_SPEED_BLACK};
 
 void init_tank(tank_t *tank) {
     tank->phys.type = PHYS_TANK;
